+++ hahomematic/caches/persistent.py
@@ -28,7 +28,7 @@ from hahomematic.support import (
     get_split_channel_address,
 )
 
-_LOGGER = logging.getLogger(__name__)
+_LOGGER = logging.getLogger("hahomematic.cache")
 
 
 class BasePersistentCache(ABC):