+++ hahomematic/central_unit.py
@@ -72,7 +72,7 @@ from hahomematic.support import (
 )
 from hahomematic.xml_rpc_proxy import XmlRpcProxy
 
-_LOGGER = logging.getLogger(__name__)
+_LOGGER = logging.getLogger("hahomematic.central")
 
 _R = TypeVar("_R")
 _T = TypeVar("_T")
@@ -249,7 +249,7 @@ class CentralUnit:
     async def start(self) -> None:
         """Start processing of the central unit."""
         if self._started:
-            _LOGGER.debug("START_: Cental %s already started", self._attr_name)
+            _LOGGER.debug("START: Cental %s already started", self._attr_name)
             return
         await self.parameter_visibility.load()
         if self.config.start_direct:
@@ -423,7 +423,7 @@ class CentralUnit:
         """De-init clients."""
         for name, client in self._clients.items():
             if await client.proxy_de_init():
-                _LOGGER.debug("STOP: Proxy de-initialized: %s", name)
+                _LOGGER.debug("DE_INIT_CLIENTS: Proxy de-initialized: %s", name)
 
     async def _init_hub(self) -> None:
         """Init the hub."""
@@ -790,7 +790,7 @@ class CentralUnit:
                     callback(value)
             except RuntimeError as rte:  # pragma: no cover
                 _LOGGER.debug(
-                    "event: RuntimeError [%s]. Failed to call callback for: %s, %s, %s",
+                    "EVENT: RuntimeError [%s]. Failed to call callback for: %s, %s, %s",
                     reduce_args(args=rte.args),
                     interface_id,
                     channel_address,
@@ -808,7 +808,7 @@ class CentralUnit:
     @callback_system_event(system_event=HmSystemEvent.LIST_DEVICES)
     def list_devices(self, interface_id: str) -> list[dict[str, Any]]:
         """Return already existing devices to CCU / Homegear."""
-        _LOGGER.debug("list_devices: interface_id = %s", interface_id)
+        _LOGGER.debug("LIST_DEVICES: interface_id = %s", interface_id)
 
         return self.device_descriptions.get_raw_device_descriptions(interface_id=interface_id)
 
@@ -838,7 +838,7 @@ class CentralUnit:
         """Remove device to central collections."""
         if device.device_address not in self._devices:
             _LOGGER.debug(
-                "remove_device: device %s not registered in central",
+                "REMOVE_DEVICE: device %s not registered in central",
                 device.device_address,
             )
             return