+++ hahomematic/platforms/entity.py
@@ -41,7 +41,7 @@ from hahomematic.platforms.support import (
     value_property,
 )
 
-_LOGGER = logging.getLogger(__name__)
+_LOGGER = logging.getLogger("hahomematic.platform")
 
 _CONFIGURABLE_CHANNEL: Final[tuple[str, ...]] = (
     "KEY_TRANSCEIVER",