+++ hahomematic/platforms/generic/entity.py
@@ -16,7 +16,7 @@ from hahomematic.exceptions import HaHomematicException
 from hahomematic.platforms import device as hmd, entity as hme
 from hahomematic.platforms.support import EntityNameData, config_property, get_entity_name
 
-_LOGGER = logging.getLogger(__name__)
+_LOGGER = logging.getLogger("hahomematic.platform")
 
 
 class GenericEntity(hme.BaseParameterEntity[hme.ParameterT, hme.InputParameterT]):