+++ hahomematic/platforms/generic/sensor.py
@@ -12,7 +12,7 @@ from hahomematic.const import HmPlatform
 from hahomematic.platforms.generic.entity import GenericEntity
 from hahomematic.platforms.support import value_property
 
-_LOGGER = logging.getLogger(__name__)
+_LOGGER = logging.getLogger("hahomematic.platform")
 
 
 class HmSensor(GenericEntity[Any, None]):