+++ hahomematic/platforms/hub/select.py
@@ -11,7 +11,7 @@ from hahomematic.const import HmPlatform
 from hahomematic.platforms.hub.entity import GenericSystemVariable
 from hahomematic.platforms.support import value_property
 
-_LOGGER = logging.getLogger(__name__)
+_LOGGER = logging.getLogger("hahomematic.platform")
 
 
 class HmSysvarSelect(GenericSystemVariable):