+++ hahomematic/platforms/hub/sensor.py
@@ -12,7 +12,7 @@ from hahomematic.const import HmPlatform, HmSysvarType
 from hahomematic.platforms.hub.entity import GenericSystemVariable
 from hahomematic.platforms.support import value_property
 
-_LOGGER = logging.getLogger(__name__)
+_LOGGER = logging.getLogger("hahomematic.platform")
 
 
 class HmSysvarSensor(GenericSystemVariable):