+++ hahomematic/platforms/support.py
@@ -23,7 +23,7 @@ from hahomematic.support import to_bool
 G = TypeVar("G")  # think about variance
 S = TypeVar("S")
 
-_LOGGER = logging.getLogger(__name__)
+_LOGGER = logging.getLogger("hahomematic.platform")
 
 # dict with binary_sensor relevant value lists and the corresponding TRUE value
 _BINARY_SENSOR_TRUE_VALUE_DICT_FOR_VALUE_LIST: Final[dict[tuple[str, ...], str]] = {